refactor: IsDeleted을 전체 도메인에서 사용할 수 있도록 위치 변경과 멤버에 IsDeleted 도입
