+++ boomerang/src/main/java/boomerang/member/domain/Member.java
@@ -1,5 +1,6 @@
 package boomerang.member.domain;
 
+import boomerang.IsDeleted;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
@@ -27,7 +28,7 @@ public class Member {
     private Long id;
 
     @Embedded
-    private Email email;
+    private String email;
 
     @Embedded
     private MemberType memberType;
@@ -47,7 +48,7 @@ public class Member {
     private SafetyScore safetyScore;
 
     @Embedded
-    private ProfileImage profileImage;
+    private String profileImage;
 
     @Embedded
     private ProgressStep progressStep;
@@ -61,12 +62,12 @@ public class Member {
     private LocalDateTime updatedAt;
 
     @Column(name = "is_deleted")
-    private boolean isDeleted;
+    private IsDeleted isDeleted;
 
     protected Member() {
     }
 
-    public Member(Email email, String nickname) {
+    public Member(String email, String nickname) {
         this.email = email;
         this.nickname = nickname;
     }
